fix(integrations): harden Copilot proxy compatibility
