--- .github/workflows/finish-pr205.yml
@@ -1,59 +1,0 @@
-name: Finish PR 205
-
-on:
-  pull_request:
-    branches: [main]
-
-permissions:
-  contents: write
-
-jobs:
-  finish:
-    if: github.head_ref == 'feat/agent-compatibility-matrix' && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 20
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: ${{ github.head_ref }}
-          fetch-depth: 0
-          token: ${{ secrets.GITHUB_TOKEN }}
-      - uses: actions/setup-python@v6
-        with:
-          python-version: '3.12'
-      - name: Apply fail-closed product repairs
-        run: python .github/finish_pr205.py
-      - name: Install focused test tools
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install -e . pytest ruff
-      - name: Compile
-        run: python -m compileall -q entroly/proxy.py entroly/cli.py entroly/session_attach.py
-      - name: Lint
-        run: python -m ruff check entroly/proxy.py entroly/cli.py entroly/session_attach.py tests/test_proxy_providers.py tests/test_agent_compatibility.py
-      - name: Focused tests
-        run: |
-          python -m pytest -q \
-            tests/test_proxy_providers.py \
-            tests/test_agent_compatibility.py \
-            tests/test_session_attach.py \
-            tests/test_cli.py \
-            tests/test_cli_audit.py
-      - name: Verify reviewable diff
-        run: |
-          test ! -e .github/copilot_composite_e2e_probe.py
-          test ! -e .github/workflows/copilot-argv-probe.yml
-          test ! -e .github/workflows/copilot-composite-e2e-probe.yml
-          test ! -e .github/workflows/copilot-composite-e2e-probe-v2.yml
-          test ! -e .github/workflows/copilot-composite-e2e-probe-v3.yml
-          ! git grep -ni -E 'kimi' -- . ':!*.lock'
-          git diff --check
-      - name: Commit product-only repairs
-        run: |
-          rm -f .github/finish_pr205.py
-          rm -f .github/workflows/finish-pr205.yml
-          git config user.name juyterman1000
-          git config user.email 208309368+juyterman1000@users.noreply.github.com
-          git add -A
-          git commit -m 'fix(integrations): harden Copilot proxy compatibility'
-          git push origin HEAD:${{ github.head_ref }}
